+++ changelog.md
@@ -12,6 +12,9 @@
 * #348 Messaging: OrderPlacedStoreOwnerNotification overwrites email account sender name with the customer's name
 * Default value for plugin description not loaded into edit popup window
 
+###Improvements###
+* #250 Implemented validation to theme configuration editing
+
 
 ##SmartStore.NET 2.0.1##